requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
httpx>=0.28,<1.0
pillow>=11.0,<12.0
pydantic-settings>=2.7,<3.0
python-multipart>=0.0.18
anthropic>=0.49,<1.0
